% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{check_install_packages}
\alias{check_install_packages}
\title{Check and Install Missing Packages}
\usage{
check_install_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names to check and install if necessary.}
}
\value{
NULL. The function is called for its side effect of installing missing packages.
}
\description{
This function checks if a series of packages are installed,
and installs those that are not installed yet. It also prints
messages to indicate which packages are already installed,
and which are being installed.
}
\examples{
\dontrun{
check_install_packages(c("dplyr", "ggplot2"))
}
}
