% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-clean.R
\name{clean_occ}
\alias{clean_occ}
\title{Clean Occurrence Data}
\usage{
clean_occ(occ_data)
}
\arguments{
\item{occ_data}{A data frame containing the occurrence data to be cleaned. It must include columns such as
\code{basisofrecord}, \code{coordinateuncertaintyinmeters}, \code{decimallongitude}, \code{decimallatitude}, and \code{countrycode}.}
}
\value{
A cleaned data frame with the same columns as the input, but with filtered and cleaned coordinates.
}
\description{
This function applies basic filters and cleans coordinates of the occurrence data. It filters for "HUMAN_OBSERVATION"
and ensures that the coordinate uncertainty is less than or equal to 10,000 meters. It also applies various cleaning
tests to remove problematic coordinates (e.g., duplicates, zeros, etc.).
}
\keyword{internal}
