% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-clean.R
\name{clean_occ_files}
\alias{clean_occ_files}
\title{Clean Occurrence Files}
\usage{
clean_occ_files(input_dir, output_dir)
}
\arguments{
\item{input_dir}{The directory containing the \code{.rds} files with the occurrence data.}

\item{output_dir}{The directory where the cleaned occurrence files will be saved.}
}
\value{
None. This function saves the cleaned occurrence data as \code{.rds} files in the specified output directory.
}
\description{
This function processes all \code{.rds} files in a specified directory, applies the cleaning function to each file,
and saves the cleaned data into an output directory. The cleaning process includes basic filtering and removing
problematic coordinates.
}
\examples{
\dontrun{
clean_occ_files("path/to/input", "path/to/output")
}
}
\seealso{
\code{\link[CoordinateCleaner]{clean_coordinates}}
}
