% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-project.R
\name{create_gitignore}
\alias{create_gitignore}
\title{Create .gitignore File}
\usage{
create_gitignore(project_path = ".", overwrite_gitignore = FALSE)
}
\arguments{
\item{project_path}{A character string specifying the path to the project root directory.
The default is the current working directory (".").}

\item{overwrite_gitignore}{A logical value indicating whether to overwrite an existing
.gitignore file in the project directory. Defaults to \code{FALSE}.}
}
\value{
A logical value (\code{TRUE}) indicating if the .gitignore file creation was successful.
}
\description{
This function creates a .gitignore file in the project's root directory based on a template.
}
\examples{
\dontrun{
# Create or overwrite .gitignore in the current directory
create_gitignore(overwrite_gitignore = TRUE)

# Create .gitignore in a specified directory without overwriting
create_gitignore("/path/to/your/project")
}
}
