% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-project.R
\name{create_quarto_yaml}
\alias{create_quarto_yaml}
\title{Create Quarto YAML File}
\usage{
create_quarto_yaml(project_path = ".", overwrite_quarto = FALSE)
}
\arguments{
\item{project_path}{A character string specifying the path to the project root directory.
The default is the current working directory (".").}

\item{overwrite_quarto}{A logical value indicating whether to overwrite an existing
_quarto.yml file in the project directory. Defaults to \code{FALSE}.}
}
\value{
A logical value (\code{TRUE}) indicating if the _quarto.yml file creation was successful.
}
\description{
This function creates a _quarto.yml file in the project's root directory based on a template.
}
\examples{
\dontrun{
# Create or overwrite _quarto.yml in the current directory
create_quarto_yaml(overwrite_quarto = TRUE)

# Create _quarto.yml in a specified directory without overwriting
create_quarto_yaml("/path/to/your/project")
}
}
