% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-project.R
\name{create_subfolders}
\alias{create_subfolders}
\title{Create Project Subfolders}
\usage{
create_subfolders(project_path = ".", subfolders = c("R", "vignettes"))
}
\arguments{
\item{project_path}{A character string specifying the path to the project root directory.
The default is the current working directory (".").}

\item{subfolders}{A character vector indicating the names of subfolders to be created within
the project directory. The default subfolders are \code{c("R", "vignettes")}.}
}
\value{
A logical value (\code{TRUE}) indicating if the subfolder creation was successful.
}
\description{
This function sets up specified subfolders within a project's root directory.
}
\examples{
\dontrun{
# Create default subfolders in the current directory
create_subfolders()

# Create specified subfolders in a specified directory
create_subfolders("/path/to/your/project", c("R", "vignettes"))
}
}
