% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo.R
\name{create_symlink_turbo}
\alias{create_symlink_turbo}
\title{Create Symbolic Link to UMich Turbo}
\usage{
create_symlink_turbo(
  project_symlink,
  turbo_target,
  turbo_volume = "seas-zhukai"
)
}
\arguments{
\item{project_symlink}{Path relative to the project root where the symbolic link will be created.}

\item{turbo_target}{Path relative to the Turbo mount point.}

\item{turbo_volume}{Name of the Turbo volume, default is 'seas-zhukai'.}
}
\value{
Logical value indicating if the symbolic link was successfully created.
}
\description{
Create a symbolic link in the project directory to UMich Turbo Research Storage.
}
\examples{
\dontrun{
create_symlink_turbo("data", "turbo-folder")
}
}
