% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-custom-download.R
\name{gbif_custom_download}
\alias{gbif_custom_download}
\title{Custom GBIF Data Download}
\usage{
gbif_custom_download(lotvs_backbone_path)
}
\arguments{
\item{lotvs_backbone_path}{The path to the \code{lotvs_backbone} RDS file containing
the usage keys.}
}
\value{
None. This function primarily initiates the download request and prints
the total occurrence count.
}
\description{
This function initiates the download of customized GBIF data. It loads the
\code{lotvs_backbone} from the provided path, splits the taxon usage keys into chunks,
calculates the total occurrence count for the taxa, and then submits the
occurrence download request for those taxa.
}
\examples{
\dontrun{
gbif_custom_download("path/to/lotvs_backbone.rds")
}
}
