% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-custom-retrieval.R
\name{gbif_custom_retrieve}
\alias{gbif_custom_retrieve}
\title{Custom GBIF Data Retrieval}
\usage{
gbif_custom_retrieve(
  taxonomy_file_path,
  data_file_path,
  species_dir_path,
  genus_dir_path
)
}
\arguments{
\item{taxonomy_file_path}{The path to the taxonomy file (RDS format) containing
the taxonomic information.}

\item{data_file_path}{The path to the directory containing the GBIF occurrence data.}

\item{species_dir_path}{The directory path to save species-level data.}

\item{genus_dir_path}{The directory path to save genus-level data.}
}
\value{
None. This function writes the filtered data into files at the specified
paths.
}
\description{
This function retrieves the GBIF data and writes it into separate files based
on taxonomic ranks (species and genus). It first reads the taxonomy data and
then filters and writes the data for species and genus into different directories.
}
\examples{
\dontrun{
gbif_custom_retrieve("path/to/taxonomy.rds", "path/to/data", 
  "path/to/species", "path/to/genus")
}
}
