% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-snapshot-download.R
\name{gbif_snapshot_download}
\alias{gbif_snapshot_download}
\title{Download and Filter GBIF Snapshot with Filtering and Partitioning}
\usage{
gbif_snapshot_download(
  bucket_fs,
  snapshot_path,
  local_save_dir,
  filter_level = c("kingdom", "phylum"),
  filter_value = c("Plantae", "Tracheophyta"),
  partition_columns = c("class", "order")
)
}
\arguments{
\item{bucket_fs}{The filesystem connection for the S3 bucket, used to access the snapshot.}

\item{snapshot_path}{The path to the GBIF snapshot on the S3 bucket.}

\item{local_save_dir}{The local directory to save the filtered snapshot.}

\item{filter_level}{A character vector of classification levels to filter the data by (e.g., "kingdom", "phylum").
Default is \code{c("kingdom", "phylum")}.}

\item{filter_value}{A character vector of values for the classification levels (e.g., "Plantae", "Tracheophyta").
Default is \code{c("Plantae", "Tracheophyta")}.}

\item{partition_columns}{A character vector specifying which columns to use for partitioning the data.
For example, \code{c("class", "order")}.}
}
\value{
None. This function writes the filtered data into files at the specified local directory.
}
\description{
This function downloads a GBIF snapshot, filters it based on the specified biological classification levels and values,
and saves the filtered data into the specified local directory in Parquet format with flexible partitioning.
}
\examples{
gbif_snapshot_download(bucket_fs, "gbif_snapshot_url", "local_dir", 
  filter_level = c("kingdom", "phylum"), filter_value = c("Plantae", "Tracheophyta"), partition_columns = c("class", "order"))
}
