% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-custom-download.R
\name{get_total_occurrence_count}
\alias{get_total_occurrence_count}
\title{Get Total Occurrence Count}
\usage{
get_total_occurrence_count(taxon_chunks)
}
\arguments{
\item{taxon_chunks}{A list of taxon key chunks to be processed.}
}
\value{
The total occurrence count.
}
\description{
This function calculates the total number of occurrences for a list of taxon keys
}
\examples{
taxon_chunks <- list(c("key1", "key2"), c("key3", "key4"))
get_total_occurrence_count(taxon_chunks)
# Returns: total occurrence count for the provided chunks
}
\keyword{internal}
