% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-package.R
\name{load_packages}
\alias{load_packages}
\title{Load Specified Packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names that need to be loaded
into the session.}
}
\value{
NULL. This function is used for its side effects of loading packages
and providing console messages.
}
\description{
This function loads a list of specified packages into the current R session.
For each package, it prints messages to inform the user about the loading
process, indicating whether loading is successful.
}
\examples{
\dontrun{
load_packages(c("dplyr", "ggplot2"))
}
}
\seealso{
\code{\link{library}}
}
