% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-snapshot-retrieval.R
\name{process_taxonomy_record}
\alias{process_taxonomy_record}
\title{Process Taxonomy Record}
\usage{
process_taxonomy_record(taxonomy_record, parquet_path, save_path)
}
\arguments{
\item{taxonomy_record}{A subset of the taxonomy list that contains information about a specific class and order.}

\item{parquet_path}{The path to the GBIF snapshot data (in Parquet format) for the specific class and order.}

\item{save_path}{The directory where the filtered species and genus data will be saved.}
}
\value{
None. This function saves species and genus occurrence data as \code{.rds} files in the specified directory.
}
\description{
This function processes the taxonomy records for a given class and order, retrieves the corresponding GBIF
occurrence data, filters the records based on species and genus, and saves the results as \code{.rds} files.
}
\keyword{internal}
