% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-snapshot-download.R
\name{setup_s3_bucket}
\alias{setup_s3_bucket}
\title{Set Up S3 Bucket Connection}
\usage{
setup_s3_bucket(bucket_name, endpoint, region, proxy = NULL)
}
\arguments{
\item{bucket_name}{The name of the S3 bucket to connect to.}

\item{endpoint}{The endpoint URL of the S3 service.}

\item{region}{The AWS region where the bucket is located.}

\item{proxy}{Optional. A list of proxy options to use for the connection. If not provided, a direct connection is used.}
}
\value{
An S3 bucket connection object.
}
\description{
This function sets up a connection to an S3 bucket, using either a proxy or direct connection, depending on the provided parameters.
}
\examples{
setup_s3_bucket(bucket_name = "gbif-open-data-us-east-1",
endpoint = "https://s3.us-east-1.amazonaws.com",
region = "us-east-1",
proxy = "http://proxy1.arc-ts.umich.edu:3128" # For using Slurm on Greatlakes)
}
