% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-custom-download.R
\name{split_into_chunks}
\alias{split_into_chunks}
\title{Split a Vector into Chunks}
\usage{
split_into_chunks(x, chunk_size)
}
\arguments{
\item{x}{A vector to be split into chunks.}

\item{chunk_size}{The size of each chunk.}
}
\value{
A list of chunks, where each chunk is a subvector of the input vector.
}
\description{
This function splits a taxon vector into smaller chunks of a specified size.
}
\examples{
split_into_chunks(1:10, 3)
# Returns: list(1:3, 4:6, 7:9, 10)
}
\keyword{internal}
