% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{update_packages}
\alias{update_packages}
\title{Update Specified Packages}
\usage{
update_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names to update.}
}
\value{
NULL. The function is called for its side effect of updating packages.
}
\description{
This function updates a specified series of packages to the latest version
on CRAN, including their dependencies.
}
\examples{
\dontrun{
update_packages(c("dplyr", "ggplot2"))
}
}
