% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-gbif-custom-retrieval.R
\name{write_chunks}
\alias{write_chunks}
\title{Write Dataset Chunks Based on Taxonomic Rank}
\usage{
write_chunks(local_df, lotvs_backbone_taxonomy, rank, dir_path, chunk_size)
}
\arguments{
\item{local_df}{The dataset to be filtered and written. This is typically a
\code{DataFrame} or a \code{dataset} from the \code{arrow} package.}

\item{lotvs_backbone_taxonomy}{A data frame containing taxonomic information
used for filtering the dataset, which includes the \code{usageKey} for species
and genus columns.}

\item{rank}{A character string, either "SPECIES" or "GENUS", indicating the rank
for filtering the data.}

\item{dir_path}{The directory where the filtered data will be saved.}

\item{chunk_size}{The number of records per chunk to be written.}
}
\value{
None. This function writes the data into files at the specified paths.
}
\description{
This function writes the dataset into separate files based on taxonomic rank
(species or genus) and saves them in the specified directories. It splits
the data into chunks for efficient processing and writing.
}
\examples{
\dontrun{
write_chunks(local_df, lotvs_backbone_taxonomy, "SPECIES", "path/to/species_dir", 100)
}
}
\keyword{internal}
