% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{check_install_packages}
\alias{check_install_packages}
\title{Check and Install Missing Packages}
\usage{
check_install_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names that need to be checked
and possibly installed.}
}
\value{
NULL. This function primarily generates side effects (installing
missing packages) and outputs messages to the console.
}
\description{
This function checks whether each package in a specified list is installed.
If any packages are missing, they are automatically installed. The function
also provides messages to indicate the status of each package, either
confirming that it's already installed or notifying the user of the
installation progress.
}
\examples{
\dontrun{
check_install_packages(c("dplyr", "ggplot2"))
}
}
\seealso{
\code{\link{install.packages}}, \code{\link{installed.packages}}
}
