% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbo.R
\name{create_symlink_turbo}
\alias{create_symlink_turbo}
\title{Create a Symbolic Link to UMich Turbo Storage}
\usage{
create_symlink_turbo(
  project_symlink,
  turbo_target,
  turbo_volume = "seas-zhukai"
)
}
\arguments{
\item{project_symlink}{A character string defining the path, relative to the project root, where the symbolic link will be created.
This must be a path within the R project directory. It cannot be a path on Turbo.}

\item{turbo_target}{A character string specifying the path, relative to the Turbo mount point, where the symbolic link should point.}

\item{turbo_volume}{A character string representing the name of the Turbo volume. The default volume is 'seas-zhukai'.}
}
\value{
A logical value indicating whether the symbolic link was successfully created (\code{TRUE}) or not (\code{FALSE}).
The function mainly operates via side effects (creating or managing symbolic links) and informs the user through messages.
}
\description{
This function generates a symbolic link within a project directory pointing to a specified location
on the University of Michigan Turbo Research Storage system. It determines the appropriate mount point
based on the operating system and creates the link accordingly. Importantly, the symbolic link must
reside within the current R project's directory. The link cannot be on Turbo because symbolic links
on Turbo differ across operating systems.
}
\examples{
\dontrun{
# Example to create a symbolic link named "data" pointing to "turbo-folder" on the Turbo storage
create_symlink_turbo("data", "turbo-folder")
}
}
\seealso{
\code{\link{file.symlink}}, \code{\link{unlink}}, \code{\link{Sys.info}}, \code{\link[R.utils]{createLink}}
}
