% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-setup.R
\name{initialize_project}
\alias{initialize_project}
\title{Initialize an R Project Structure}
\usage{
initialize_project(
  project_path = ".",
  subfolders = c("R", "tests", "vignettes"),
  overwrite_gitignore = FALSE
)
}
\arguments{
\item{project_path}{A character string specifying the path to the project root directory.
The default is the current working directory (".").}

\item{subfolders}{A character vector indicating the names of subfolders to be created within
the project directory. The default subfolders are \code{c("R", "tests", "vignettes")}.}

\item{overwrite_gitignore}{A logical value indicating whether to overwrite an existing
.gitignore file in the project directory. Defaults to \code{FALSE}.}
}
\value{
A logical value (\code{TRUE}) indicating if the initialization was successful.
The function is primarily used for its side-effect of creating folders and a .gitignore file.
}
\description{
This function sets up a standardized file structure for an R project within a specified directory.
It creates specified subfolders and a .gitignore file to manage unwanted files in version control.
}
\examples{
\dontrun{
# Initialize project folders and .gitignore in the current directory
initialize_project()

# Initialize project folders and .gitignore in a specified directory
initialize_project("/path/to/your/project")

# Initialize project folders with custom subfolders and overwrite existing .gitignore file
initialize_project(
  project_path = "/path/to/your/project",
  subfolders = c("R", "data", "docs"),
  overwrite_gitignore = TRUE
)
}
}
\seealso{
\code{\link{dir.create}}, \code{\link{file.path}}, \code{\link{writeLines}}
}
