% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-utils.R
\name{load_packages}
\alias{load_packages}
\title{Load Specified Packages}
\usage{
load_packages(packages)
}
\arguments{
\item{packages}{A character vector of package names to load.}
}
\value{
NULL. The function is called for its side effect of loading packages.
}
\description{
This function loads a series of packages into the R session.
It also prints messages to indicate which packages are being loaded.
}
\examples{
\dontrun{
load_packages(c("dplyr", "ggplot2"))
}
}
